don't manage reindex queue in tests
The recipe reindexing queue (as well as the index itself) is specific
to PostgreSQL, so disable it in the test profile. It has its own unit
tests, so not as silly as it sounds.
